Apply perf v5 and watcher fixes

+++ src/main/java/net/minecraft/entity/titan/EntityGargoyleTitanFireball.java
@@ -163,8 +163,8 @@ extends EntityFireball {
     }
 
     protected void entityInit() {
-        this.dataWatcher.addObject(10, (Object)0);
-        this.dataWatcher.addObject(11, (Object)0);
+        this.dataWatcher.addObject(10, (Object)Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(11, (Object)Byte.valueOf((byte)0));
     }
 
     public int getBlockType() {
