+++ src/main/java/net/minecraft/entity/titan/EntitySpiderTitan.java
@@ -204,7 +204,7 @@ IEntityMultiPartTitan {
     protected void entityInit() {
         super.entityInit();
         this.dataWatcher.addObject(16, (Object)Byte.valueOf((byte)0));
-        this.dataWatcher.addObject(17, (Object)0);
+        this.dataWatcher.addObject(17, (Object)Byte.valueOf((byte)0));
     }
 
     public int getBonusID() {
