+++ src/main/java/net/minecraft/entity/titanminion/EntityGhastGuard.java
@@ -339,7 +339,7 @@ IMinion {
 
     protected void entityInit() {
         super.entityInit();
-        this.dataWatcher.addObject(16, (Object)0);
+        this.dataWatcher.addObject(16, (Object)Byte.valueOf((byte)0));
     }
 
     protected void applyEntityAttributes() {
